+++ CHANGELOG.md
@@ -22,6 +22,14 @@ You can easily find changes by searching for `"### Change"` in this markdown fil
 # Releases
 Test/interim releases are not documented here.
 
+## [0.6.?]
+### Added
+- Add `c_array_mem<T>`
+    - https://github.com/fin-language/fin/issues/87
+
+### Changed (minor)
+- math mode defaults to `unsafe mode` for now.
+
 ---
 
 ## [0.6.0]